+++ espnet2/train/class_choices.py
@@ -1,0 +1,3 @@
+from __future__ import annotations
+
+
